requirement.txt
absl-py==0.11.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.10.0
anaconda-project==0.8.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.2.0
Automat==20.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
bs4==0.0.1
cachetools==4.2.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.9.2
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.3
distributed==2.5.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-headers==1.0.2
fake-useragent==0.1.11
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
fsspec==0.5.2
future==0.17.1
gast==0.3.3
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
graphviz==0.14.1
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==20.0.1
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itemadapter==0.1.1
itemloaders==1.0.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jieba==0.41
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nlp==0.4.0
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.19.5
numpydoc==0.9.1
oauthlib==3.1.0
objgraph==3.4.1
olefile==0.46
opencc-python-reimplemented==0.1.5
openpyxl==3.0.0
opt-einsum==3.3.0
packaging==19.2
paddlepaddle-tiny==1.6.1
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.6.0
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pdfminer==20191125
pdfminer.six==20201018
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pipenv==2020.6.2
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.12.2
psutil==5.6.3
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.9.7
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==2.0.2
pylint==2.4.2
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-Levenshtein==0.12.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
queuelib==1.5.0
regex==2020.11.13
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.7
ruamel-yaml==0.15.46
sacremoses==0.0.43
scikit-image==0.15.0
scikit-learn==0.23.1
scipy==1.3.1
Scrapy==2.4.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
shap==0.39.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
slicer==0.0.7
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toolz==0.10.0
torch==1.8.0
torchvision==0.9.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
transformers==4.4.2
Twisted==20.3.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
ujson==4.0.2
unicodecsv==0.14.1
urllib3==1.24.2
virtualenv==16.7.9
virtualenv-clone==0.5.3
w3lib==1.22.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.12.1
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwings==0.15.10
xlwt==1.3.0
xmltodict==0.12.0
xxhash==2.0.0
zict==1.0.0
zipp==0.6.0
zope.interface==5.1.2
